requirements.txt
blinker==1.9.0
click==8.2.0
flask==3.1.1
flask-wtf==1.2.2
itsdangerous==2.2.0
jinja2==3.1.6
markupsafe==3.0.2
python-dotenv==1.1.0
werkzeug==3.1.3
wtforms==3.2.1
